+++ src/main/java/de/samply/frontend/Action.java
@@ -1,7 +1,8 @@
 package de.samply.frontend;
 
 public record Action(
-        String action,
-        String path
+        String path,
+        String method,
+        String[] params
 ) {
 }