+++ src/main/java/de/samply/frontend/ModuleActionsPackage.java
@@ -2,19 +2,20 @@ package de.samply.frontend;
 
 import lombok.Getter;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
 @Getter
 public class ModuleActionsPackage {
     private String module;
-    private List<Action> actions = new ArrayList<>();
+    private Map<String, Action> actions = new HashMap<>();
 
     public void setModule(String module) {
         this.module = module;
     }
 
-    public void addAction(Action action) {
-        actions.add(action);
+    public void addAction(String actionName, Action action) {
+        actions.put(actionName, action);
     }
+
 }