+++ src/main/java/de/samply/utils/AspectUtils.java
@@ -7,10 +7,15 @@ import de.samply.db.model.Project;
 import de.samply.db.repository.ProjectRepository;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.reflect.MethodSignature;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.Optional;
+import java.util.Set;
 
 public class AspectUtils {
 
@@ -53,5 +58,31 @@ public class AspectUtils {
         return (projectCode.isPresent()) ? projectRepository.findByCode(projectCode.get()) : Optional.empty();
     }
 
+    public static String[] fetchRequestParamNames(Method method) {
+        Set<String> result = new HashSet<>();
+        Arrays.stream(method.getParameters()).forEach(parameter -> Arrays.stream(parameter.getAnnotations())
+                .filter(annotation -> annotation.annotationType() == RequestParam.class)
+                .map(annotation -> (RequestParam) annotation)
+                .forEach(requestParam -> result.add(requestParam.name())));
+        return result.toArray(String[]::new);
+    }
+
+    public static Optional<String> fetchHttpMethod(Method method) {
+        RequestMapping annotation = method.getAnnotation(RequestMapping.class);
+        if (annotation == null) {
+            for (Annotation methodAnnotation : method.getDeclaredAnnotations()) {
+                for (Annotation methodAnnotationAnnotation : methodAnnotation.annotationType().getDeclaredAnnotations()) {
+                    if (methodAnnotationAnnotation.annotationType() == RequestMapping.class) {
+                        return fetchHttpMetod((RequestMapping) methodAnnotationAnnotation);
+                    }
+                }
+            }
+        }
+        return fetchHttpMetod(annotation);
+    }
+
+    private static Optional<String> fetchHttpMetod(RequestMapping requestMapping) {
+        return Optional.of(requestMapping.method()[0].name());
+    }
 
 }